add current updates to update mutation (based on review)
