+++ app/graphql/schema/mutations/index.js
@@ -1,7 +1,7 @@
-const { createCustomerQueryResponse } = require('./create-customer-query-response')
+const { updateCustomerQueryTicket } = require('./update-customer-query-ticket')
 const { createCustomerQueryTicket } = require('./create-customer-query-ticket')
 
 module.exports = {
-  createCustomerQueryResponse,
+  updateCustomerQueryTicket,
   createCustomerQueryTicket
 }
